Posting a solution, dialog boxes
The code for posting a solution is up and running. I created a couple
schemas for dialogue boxes, we should all use these for consistency.
Cleaned up some random other files. NOTE: we still need to contact
question services on startup and on exit.. i did not add code for this
yet, but will on next commit (posting questions)

Conflicts:
	InterviewAnnihilator/res/layout/activity_solution.xml
	InterviewAnnihilator/res/values/strings.xml
	InterviewAnnihilator/src/com/huskysoft/interviewannihilator/runtime/FetchQuestionsTask.java
	InterviewAnnihilator/src/com/huskysoft/interviewannihilator/runtime/FetchSolutionsTask.java
	InterviewAnnihilator/src/com/huskysoft/interviewannihilator/ui/MainActivity.java
	InterviewAnnihilator/src/com/huskysoft/interviewannihilator/ui/PostSolutionActivity.java
	InterviewAnnihilator/src/com/huskysoft/interviewannihilator/ui/QuestionActivity.java